+++ zoomimage-view/src/main/kotlin/com/github/panpf/zoomimage/view/zoom/ZoomableEngine.kt
@@ -106,7 +106,8 @@ class ZoomableEngine constructor(val logger: Logger, val view: View) {
         MutableStateFlow(zoomableCore.contentSize)
 
     /**
-     * The original size of the content, it is usually set by [SubsamplingEngine] after parsing the original size of the image
+     * The original size of the content, it is usually set by [SubsamplingEngine] after parsing the original size of the image.
+     * If not empty, it means that the subsampling function has been enabled
      */
     val contentOriginSizeState: MutableStateFlow<IntSizeCompat> =
         MutableStateFlow(zoomableCore.contentOriginSize)